feat(relay-compiler): version 6 support
