+++ types/relay-compiler/index.d.ts
@@ -1,16 +1,16 @@
-// Type definitions for relay-compiler 5.0
+// Type definitions for relay-compiler 6.0
 // Project: https://relay.dev
 // Definitions by: n1ru4l <https://github.com/n1ru4l>
 // Definitions: https://github.com/DefinitelyTyped/DefinitelyTyped
 // TypeScript Version: 3.0
 
-import { GraphQLCompilerContext } from './lib/GraphQLCompilerContext';
-import * as ASTConvert from './lib/ASTConvert';
+import { GraphQLCompilerContext } from './lib/core/GraphQLCompilerContext';
+import * as ASTConvert from './lib/core/ASTConvert';
 
-import * as Parser from './lib/RelayParser';
-import * as Printer from './lib/GraphQLIRPrinter';
-import ConsoleReporter = require('./lib/GraphQLConsoleReporter');
-import MultiReporter = require('./lib/GraphQLMultiReporter');
+import * as Parser from './lib/core/RelayParser';
+import * as Printer from './lib/core/GraphQLIRPrinter';
+import ConsoleReporter = require('./lib/reporters/GraphQLConsoleReporter');
+import MultiReporter = require('./lib/reporters/GraphQLMultiReporter');
 
 declare var transformASTSchema: typeof ASTConvert.transformASTSchema;
