+++ types/relay-compiler/lib/reporters/GraphQLConsoleReporter.d.ts
