+++ types/relay-compiler/lib/reporters/GraphQLMultiReporter.d.ts
