+++ types/relay-compiler/lib/reporters/GraphQLReporter.d.ts
