+++ types/relay-compiler/lib/transforms/InlineFragmentsTransform.d.ts
@@ -1,4 +1,4 @@
-import { GraphQLCompilerContext } from '../';
+import { GraphQLCompilerContext } from '../core/GraphQLCompilerContext';
 
 export function transform(
   context: GraphQLCompilerContext,
