+++ types/relay-compiler/tsconfig.json
@@ -20,9 +20,9 @@
     "files": [
         "index.d.ts",
         "relay-compiler-tests.ts",
-        "lib/FlattenTransform.d.ts",
-        "lib/InlineFragmentsTransform.d.ts",
-        "lib/RelayApplyFragmentArgumentTransform.d.ts",
-        "lib/SkipRedundantNodesTransform.d.ts"
+        "lib/transforms/FlattenTransform.d.ts",
+        "lib/transforms/InlineFragmentsTransform.d.ts",
+        "lib/transforms/RelayApplyFragmentArgumentTransform.d.ts",
+        "lib/transforms/SkipRedundantNodesTransform.d.ts"
     ]
 }
